[xyd/env] trajectory tracking
- add environment for trajectory tracking
  - build_space/reset/step
  - unit test
- manager class
  - DynamicsModelManager
  - ObservationManager
  - ReferenceLineManager
- add registry
- add primitives for geometry (coordinate transform)
- add config systems
  - primitives for config
  - sample config
- update unit test
  - add execution section
    - if __name__ == '__main__'
- misc update
  - DTYPE move to np.float64
  - add frozendict
  - add BodyState
  - add `override` and require Python>=3.12

+++ simulator/__init__.py
@@ -1,3 +1,3 @@
 import numpy as np
 
-DTYPE = np.float32  # consistent with protobuf
+DTYPE = np.float64  # consistent with protobuf
